Clean up _Utils.kt into multiple files with purposes
